requirements.txt
google-genai
PyYAML
Pillow
rich